DiscreteFiltering

Documentation for DiscreteFiltering.

DiscreteFiltering.DiscreteFiltering

DiscreteFiltering.S

DiscreteFiltering.S!

DiscreteFiltering.circulant

DiscreteFiltering.create_data_dns

DiscreteFiltering.create_data_exact

DiscreteFiltering.create_data_filtered

DiscreteFiltering.create_gaussian

DiscreteFiltering.create_loss

DiscreteFiltering.create_tophat

DiscreteFiltering.filter_matrix

DiscreteFiltering.fit_intrusive

DiscreteFiltering.plotmat

DiscreteFiltering.relerr

DiscreteFiltering.relerrs

DiscreteFiltering.spectral_relerr

DiscreteFiltering.DiscreteFiltering — Module

Discrete filtering toolbox

DiscreteFiltering.S! — Method

S!(A, u, t; kwargs...)

Solve ODE for given operator and IC (mutating form, not differentiable).

DiscreteFiltering.S — Method

S(A, u₀, t; kwargs...)

Solve ODE for given operator and IC. This form is differentiable.

DiscreteFiltering.circulant — Method

circulant(n, inds, stencil)

Create circulant SparseMatrixCSC.

DiscreteFiltering.create_data_dns — Method

create_data_dns(A, c, ξ, t)

Create DNS solution (numerical approximation).

DiscreteFiltering.create_data_exact — Method

create_data_exact(c, ξ, t)

Create exact unfiltered solutions.

DiscreteFiltering.create_data_filtered — Method

create_data_filtered(W, A, dns)

Create discrete filtered solution from DNS

DiscreteFiltering.create_gaussian — Method

create_gaussian(h)

Create Gaussian filter with filter radius h.

The kernel is given by

\[ G(x, \xi) = \sqrt{\frac{3}{2 \pi h^2(x)}} \mathrm{e}^{-\frac{3 (\xi - x)^2}{2 h^2(x)}},\]

and the local transfer function by

\[ \hat{G}_k(x) = \mathrm{e}^{-\frac{2 \pi^2}{3} k^2 h^2(x)}.\]

DiscreteFiltering.create_loss — Method

Create loss function on dataset.

DiscreteFiltering.create_tophat — Method

create_tophat(h)

Create top-hat filter with filter radius h.

The kernel is given by

\[ G(x, \xi) = \begin{cases} \frac{1}{2h(x)} \quad & |\xi - x| \leq h(x), \\ 0 \quad & \text{otherwise}, \end{cases}\]

and the local transfer function by

\[ \hat{G}_k(x) = \frac{\sin \left ( 2 \pi k h(x) \right )}{2 \pi k h(x)}.\]

DiscreteFiltering.filter_matrix — Method

filter_matrix(F, x, ξ)

Create filter matrix with filter F from fine grid ξ to coarse grid x.

DiscreteFiltering.fit_intrusive — Method

fit_intrusive(
    A_ref,
    uₜ,
    t;
    α = 0.001,
    β₁ = 0.9,
    β₂ = 0.999,
    ϵ = 1e-8,
    λ = 1e-10,
    nbatch = 10,
    niter = 100,
    initial = nothing,
    testloss = A -> nothing,
    ntestloss = 10,
    ntime = 10,
    ntimebatch = 10,
    doplot = true,
    kwargs...,
)

Fit operator to data intrusively (trough the ODE solver) using the ADAM optimizer.

DiscreteFiltering.plotmat — Function

plotmat(A; kwargs...)

Plot matrix.

DiscreteFiltering.relerr — Method

relerr(A, uₜ, t; kwargs...)

Relative error, averaged over time and data samples.

DiscreteFiltering.relerrs — Method

relerrs(A, uₜ, t; kwargs...)

Time dependent relative errors, averaged over data samples.

DiscreteFiltering.spectral_relerr — Method

spectral_relerr(A, uₜ, t; kwargs...)

Time dependent spectral relative errors, averaged over data samples.