DiscreteFiltering

Documentation for DiscreteFiltering.

DiscreteFiltering.DiscreteFiltering

DiscreteFiltering.AbstractIntervalDomain

DiscreteFiltering.AdvectionEquation

DiscreteFiltering.BurgersEquation

DiscreteFiltering.ClosedIntervalDomain

DiscreteFiltering.ConvolutionalFilter

DiscreteFiltering.DiffusionEquation

DiscreteFiltering.Domain

DiscreteFiltering.Equation

DiscreteFiltering.Filter

DiscreteFiltering.IdentityFilter

DiscreteFiltering.PeriodicIntervalDomain

DiscreteFiltering.TopHatFilter

DiscreteFiltering.GaussianFilter

DiscreteFiltering.advection_matrix

DiscreteFiltering.apply_filter

DiscreteFiltering.apply_filter_extend

DiscreteFiltering.apply_filter_extend

DiscreteFiltering.diffusion_matrix

DiscreteFiltering.discretize

DiscreteFiltering.filter_matrix

DiscreteFiltering.filter_matrix_meshwidth

DiscreteFiltering.interpolation_matrix

DiscreteFiltering.reconstruction_matrix

DiscreteFiltering.reconstruction_matrix_meshwidth

DiscreteFiltering.ridge

DiscreteFiltering.solve

DiscreteFiltering.solve

DiscreteFiltering.solve

DiscreteFiltering.solve

DiscreteFiltering.solve_adbc

DiscreteFiltering.solve_burgers

DiscreteFiltering.DiscreteFiltering — Module

Discrete filtering toolbox

DiscreteFiltering.AbstractIntervalDomain — Type

ClosedIntervalDomain

Abstract type for interval domains.

DiscreteFiltering.AdvectionEquation — Type

AdvectionEquation(domain, filter = IdentityFilter())

Filtered advection equation.

DiscreteFiltering.BurgersEquation — Type

BurgersEquation(domain, filter = IdentityFilter())

Filtered Burgers equation.

DiscreteFiltering.ClosedIntervalDomain — Type

ClosedIntervalDomain(left, right)

Interval domain.

DiscreteFiltering.ConvolutionalFilter — Type

ConvolutionalFilter(kernel)

Convolutional filter, parameterized by a filter kernel.

DiscreteFiltering.DiffusionEquation — Type

DiffusionEquation(
    domain,
    filter = IdentityFilter(),
    f = nothing,
    g_a = nothing,
    g_b = nothing
)

Filtered diffusion equation.

DiscreteFiltering.Domain — Type

Domain

Abstract type for different domains.

DiscreteFiltering.Equation — Type

Abstract equation.

DiscreteFiltering.Filter — Type

Abstract continuous filter.

DiscreteFiltering.IdentityFilter — Type

IdentityFilter()

Identity filter, which does not filter.

DiscreteFiltering.PeriodicIntervalDomain — Type

PeriodicIntervalDomain(left, right)

Periodic interval domain.

DiscreteFiltering.TopHatFilter — Type

TopHatFilter(width)

Top hat filter, parameterized by a variable filter width.

DiscreteFiltering.GaussianFilter — Method

GaussianFilter(h, σ) -> ConvolutionalFilter GaussianFilter(σ) -> ConvolutionalFilter

Create Gaussian ConvolutionalFilter with domain width 2h and variance σ^2.

DiscreteFiltering.advection_matrix — Function

advection_matrix(domain, n)

Assemble discrete advection matrix.

DiscreteFiltering.apply_filter — Function

apply_filter(u, filter, domain)

Apply filter to a spatial function u defined on domain.

DiscreteFiltering.apply_filter_extend — Method

apply_filter_extend(u, filter, domain)

Extend function beyond beyond boundaries before filtering.

DiscreteFiltering.apply_filter_extend — Method

apply_filter_extend(u, filter, domain)

Extend function beyond beyond boundaries before filtering.

DiscreteFiltering.diffusion_matrix — Function

diffusion_matrix(domain, n)

Assemble discrete diffusion matrix.

DiscreteFiltering.discretize — Function

discretize(domain, N)

Discretize domain with N points.

DiscreteFiltering.filter_matrix — Function

filter_matrix(f, domain, M, N)

Assemble discrete filtering matrix from a continuous filter f.

DiscreteFiltering.filter_matrix_meshwidth — Method

filtermatrixmeshwidth(f, domain, N)

Assemble discrete filtering matrix from a continuous filter f width constant width $h(x) = \Delta x / 2$.

DiscreteFiltering.interpolation_matrix — Method

interpolation_matrix(x, ξ)

Create interpolation matrix from grid ξ to grid x.

DiscreteFiltering.reconstruction_matrix — Function

reconstruction_matrix(filter, domain, M, N)

Approximate R of size N × M, the inverse of the discrete filtering matrix W of size M × N.

DiscreteFiltering.reconstruction_matrix_meshwidth — Method

reconstructionmatrixmeshwidth(f, domain, N)

Assemble inverse discrete filtering matrix from a continuous filter f width constant width $h(x) = \Delta x / 2$.

DiscreteFiltering.ridge — Function

ridge(A, b, λ = 0)

Compute the solution to min ||Ax - b||² + λ||x||².

DiscreteFiltering.solve — Function

solve

Solve equation.

DiscreteFiltering.solve — Method

solve(
    equation::AdvectionEquation{PeriodicIntervalDomain{T},F},
    u,
    tlist,
    M,
    N;
    method = "filterfirst",
    subspacedim = 10,
    solver = QNDF(),
    reltol = 1e-4,
    abstol = 1e-6,
) where {T,F}

Solve equation from tlist[1] to tlist[2] with initial conditions u and a discretization of M and N points. If method is "filterfirst", the equation is filtered then discretized. If method is "discretizefirst", the equation is discretized then filtered. The parameter subspacedim controls the accuracy of the linear exponential timestepping.

DiscreteFiltering.solve — Method

solve(
    equation::DiffusionEquation{ClosedIntervalDomain{T},F},
    u,
    tlist,
    M,
    N;
    method = "filterfirst",
    boundary_conditions = "exact",
    solver = QNDF(),
    abstol = 1e-4,
    reltol = 1e-3,
    degmax = 10,
    λ = 1e-6
) where {T,F}

Solve equation from tlist[1] to tlist[2] with initial conditions u and a discretization of M and N points. If method is "filterfirst", the equation is filtered then discretized. If method is "discretizefirst", the equation is discretized then filtered.

DiscreteFiltering.solve — Method

solve(
    equation::DiffusionEquation{PeriodicIntervalDomain{T},F},
    u,
    tlist,
    M,
    N;
    method = "filterfirst",
    solver = QNDF(),
    abstol = 1e-4,
    reltol = 1e-3,
    degmax = 10,
    λ = 1e-6
) where {T,F}

Solve equation from tlist[1] to tlist[2] with initial conditions u and a discretization of M and N points. If method is "filterfirst", the equation is filtered then discretized. If method is "discretizefirst", the equation is discretized then filtered.

DiscreteFiltering.solve_adbc — Method

solve_adbc(equation, u, tlist, n, dt = (tlist[2] - tlist[1]) / 200)

Solve filtered equation using approximate deconvolution boundary conditions (ADBC). This approach was proposed in [Borggaard 2006].

J. Borggaard, T. Iliescu, Approximate deconvolution boundary conditions for large eddy simulation, Applied Mathematics Letters 19 (8) (2006) 735–740. doi: https://doi.org/10.1016/j.aml.2005.08.022. URL: https://www.sciencedirect.com/science/article/pii/S0893965905003319

DiscreteFiltering.solve_burgers — Method

solve_burgers()

Solve Burgers equation.