DiscreteFiltering

Documentation for DiscreteFiltering.

DiscreteFiltering.DiscreteFiltering

DiscreteFiltering.AbstractIntervalDomain

DiscreteFiltering.ClosedIntervalDomain

DiscreteFiltering.ConvolutionalFilter

DiscreteFiltering.Domain

DiscreteFiltering.Filter

DiscreteFiltering.PeriodicIntervalDomain

DiscreteFiltering.TopHatFilter

DiscreteFiltering.advection_matrix

DiscreteFiltering.diffusion_matrix

DiscreteFiltering.discretize_uniform

DiscreteFiltering.filter_matrix

DiscreteFiltering.inverse_filter_matrix

DiscreteFiltering.DiscreteFiltering — Module

Discrete filtering toolbox

DiscreteFiltering.AbstractIntervalDomain — Type

ClosedIntervalDomain

Abstract type for interval domains.

DiscreteFiltering.ClosedIntervalDomain — Type

ClosedIntervalDomain(left, right)

Interval domain.

DiscreteFiltering.ConvolutionalFilter — Type

ConvolutionalFilter(kernel)

Convolutional filter, parameterized by a filter kernel.

DiscreteFiltering.Domain — Type

Domain

Abstract type for different domains.

DiscreteFiltering.Filter — Type

Abstract continuous filter.

DiscreteFiltering.PeriodicIntervalDomain — Type

PeriodicIntervalDomain(left, right)

Periodic interval domain.

DiscreteFiltering.TopHatFilter — Type

TopHatFilter(width)

Top hat filter, parameterized by a variable filter width.

DiscreteFiltering.advection_matrix — Method

advection_matrix(domain, n)

Assemble discrete advection matrix.

DiscreteFiltering.diffusion_matrix — Method

diffusion_matrix(n)

Assemble discrete diffusion matrix.

DiscreteFiltering.discretize_uniform — Method

discretize(domain)

Discretize domain.

DiscreteFiltering.filter_matrix — Method

filter_matrix(f, x[, ε])

Assemble discrete filtering matrix from a continuous filter f.

DiscreteFiltering.inverse_filter_matrix — Method

inverse_filter_matrix(f, domain, n)

Approximate inverse of discrete filtering matrix, given filter f.