DiscreteFiltering

Documentation for DiscreteFiltering.

DiscreteFiltering.DiscreteFiltering

DiscreteFiltering.AbstractIntervalDomain

DiscreteFiltering.AdvectionEquation

DiscreteFiltering.BurgersEquation

DiscreteFiltering.ClosedIntervalDomain

DiscreteFiltering.ConvolutionalFilter

DiscreteFiltering.DiffusionEquation

DiscreteFiltering.Domain

DiscreteFiltering.Equation

DiscreteFiltering.Filter

DiscreteFiltering.IdentityFilter

DiscreteFiltering.PeriodicIntervalDomain

DiscreteFiltering.TopHatFilter

DiscreteFiltering.GaussianFilter

DiscreteFiltering.advection_matrix

DiscreteFiltering.apply_filter

DiscreteFiltering.apply_filter_extend

DiscreteFiltering.apply_filter_extend

DiscreteFiltering.diffusion_matrix

DiscreteFiltering.discretize

DiscreteFiltering.filter_matrix

DiscreteFiltering.filter_matrix_meshwidth

DiscreteFiltering.reconstruction_matrix

DiscreteFiltering.reconstruction_matrix_meshwidth

DiscreteFiltering.solve

DiscreteFiltering.solve_adbc

DiscreteFiltering.solve_burgers

DiscreteFiltering.DiscreteFiltering — Module

Discrete filtering toolbox

DiscreteFiltering.AbstractIntervalDomain — Type

ClosedIntervalDomain

Abstract type for interval domains.

DiscreteFiltering.AdvectionEquation — Type

AdvectionEquation(domain, filter = IdentityFilter())

Filtered advection equation.

DiscreteFiltering.BurgersEquation — Type

BurgersEquation(domain, filter = IdentityFilter())

Filtered Burgers equation.

DiscreteFiltering.ClosedIntervalDomain — Type

ClosedIntervalDomain(left, right)

Interval domain.

DiscreteFiltering.ConvolutionalFilter — Type

ConvolutionalFilter(kernel)

Convolutional filter, parameterized by a filter kernel.

DiscreteFiltering.DiffusionEquation — Type

DiffusionEquation(
    domain,
    filter = IdentityFilter(),
    f = nothing,
    g_a = nothing,
    g_b = nothing
)

Filtered diffusion equation.

DiscreteFiltering.Domain — Type

Domain

Abstract type for different domains.

DiscreteFiltering.Equation — Type

Abstract equation.

DiscreteFiltering.Filter — Type

Abstract continuous filter.

DiscreteFiltering.IdentityFilter — Type

IdentityFilter()

Identity filter, which does not filter.

DiscreteFiltering.PeriodicIntervalDomain — Type

PeriodicIntervalDomain(left, right)

Periodic interval domain.

DiscreteFiltering.TopHatFilter — Type

TopHatFilter(width)

Top hat filter, parameterized by a variable filter width.

DiscreteFiltering.GaussianFilter — Method

GaussianFilter(h, σ) -> ConvolutionalFilter GaussianFilter(σ) -> ConvolutionalFilter

Create Gaussian ConvolutionalFilter with domain width 2h and variance σ^2.

DiscreteFiltering.advection_matrix — Method

advection_matrix(domain, n)

Assemble discrete advection matrix.

DiscreteFiltering.apply_filter — Method

apply_filter(u, filter, domain)

Apply filter to a spatial function u defined on domain.

DiscreteFiltering.apply_filter_extend — Method

apply_filter_extend(u, filter, domain)

Extend function beyond beyond boundaries before filtering.

DiscreteFiltering.apply_filter_extend — Method

apply_filter_extend(u, filter, domain)

Extend function beyond beyond boundaries before filtering.

DiscreteFiltering.diffusion_matrix — Method

diffusion_matrix(domain, n)

Assemble discrete diffusion matrix.

DiscreteFiltering.discretize — Method

discretize(domain, n)

Discretize domain with n points.

DiscreteFiltering.filter_matrix — Method

filter_matrix(f, domain, n)

Assemble discrete filtering matrix from a continuous filter f.

DiscreteFiltering.filter_matrix_meshwidth — Method

filter_matrix_meshwidth(f, domain, n)

Assemble discrete filtering matrix from a continuous filter f width constant width $h(x) = \Delta x / 2$.

DiscreteFiltering.reconstruction_matrix — Method

reconstruction_matrix(f, domain, n)

Approximate inverse of discrete filtering matrix, given filter f.

DiscreteFiltering.reconstruction_matrix_meshwidth — Method

reconstruction_matrix_meshwidth(f, domain, n)

Assemble inverse discrete filtering matrix from a continuous filter f width constant width $h(x) = \Delta x / 2$.

DiscreteFiltering.solve — Method

solve(equation, u, tlist, n; method = "filterfirst")

Solve equation from tlist[1] to tlist[2] with initial conditions u and a discretization of n points. If method is "filterfirst", the equation is filtered then discretized. If method is "discretizefirst", the equation is discretized then filtered.

DiscreteFiltering.solve_adbc — Method

solve_adbc(equation, u, tlist, n, dt = (tlist[2] - tlist[1]) / 200)

Solve filtered equation using approximate deconvolution boundary conditions (ADBC). This approach was proposed in [Borggaard 2006].

J. Borggaard, T. Iliescu, Approximate deconvolution boundary conditions for large eddy simulation, Applied Mathematics Letters 19 (8) (2006) 735–740. doi: https://doi.org/10.1016/j.aml.2005.08.022. URL: https://www.sciencedirect.com/science/article/pii/S0893965905003319

DiscreteFiltering.solve_burgers — Method

solve_burgers()

Solve Burgers equation.